Fixed some SEL problems.
git-svn-id: file:///home/cminyard/t2/svn/openipmi/trunk@1616 a860ddb0-99b4-4eb6-8282-105f033f2ff5

+++ lib/normal_fru.c
@@ -67,7 +67,7 @@ typedef struct ipmi_fru_record_s ipmi_fru_record_t;
 typedef struct fru_string_s
 {
     enum ipmi_str_type_e type;
-    unsigned short       length;
+    unsigned int         length;
     char                 *str;
 
     /* The raw offset from the start of the area, and the raw length
